updated project

+++ src/main/java/me/eiad/vromgame/core/Track.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import me.eiad.vromgame.exeptions.LengthShouldBePositiveNumber;
 @Getter
 public class Track {
-    private final int length;
+    private final double length;
 
     public Track(int length) {
         if (length < 1) {
